+++ android/app/src/main/java/software/thinkdigital/tavern/Application.java
@@ -1,0 +1,20 @@
+package software.thinkdigital.tavern;
+
+import com.transistorsoft.flutter.backgroundfetch.BackgroundFetchPlugin;
+
+import io.flutter.app.FlutterApplication;
+import io.flutter.plugin.common.PluginRegistry;
+import io.flutter.plugins.GeneratedPluginRegistrant;
+
+public class Application extends FlutterApplication implements PluginRegistry.PluginRegistrantCallback {
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        BackgroundFetchPlugin.setPluginRegistrant(this);
+    }
+
+    @Override
+    public void registerWith(PluginRegistry registry) {
+        GeneratedPluginRegistrant.registerWith(registry);
+    }
+}