+++ CHANGELOG.md
@@ -8,11 +8,20 @@ Fix error on initial launch and missing Tavern logo.
 
 ## v1.2.0 - 1.2.2
 Fix Title overflow on PackageDetailsScreen. 
+
 Fix Content errors on AnalysisTab.
+
 Add support for Dart library packages which have no package page
+
 Update Theme to be more consistent.
-Remove Score indicator from packages with no score
+
+Remove Score indicator from packages with no score.
 Closes #49
 
 Fix: Swiping on the pageView does not update the index on the bottom navigation bar
 Closes #45.
+
+## v1.3.0 
+Add support for background fetch with subscriptions for Android and iOS.
+
+Next up: Notifications